:art: change register to use route nesting
