+++ src/App.tsx
@@ -36,8 +36,10 @@ import NotFoundPage from "./pages/not-found/NotFoundPage";
 import { BackdError } from "./app/errors";
 import { useIsLive } from "./app/hooks/use-is-live";
 import ActionsPage from "./pages/actions/ActionsPage";
-import RegisterAction from "./pages/actions/RegisterAction";
+import ActionRegister from "./pages/actions/register/ActionRegister";
 import ActionsIndex from "./pages/actions/ActionsIndex";
+import ActionRegisterIndex from "./pages/actions/register/ActionRegisterIndex";
+import RegisterTopup from "./pages/actions/lending/RegisterTopup";
 
 const Background = styled.div`
   background: radial-gradient(rgba(11, 3, 60, 0.2), rgba(10, 5, 38, 0.3));
@@ -117,7 +119,10 @@ const App = (): JSX.Element => {
                       <Route path="/pool/:poolName" element={<PoolPage />} />
                       <Route path="/pools" element={<PoolsPage />} />
                       <Route path="/actions" element={<ActionsPage />}>
-                        <Route path="register" element={<RegisterAction />} />
+                        <Route path="register" element={<ActionRegister />}>
+                          <Route path="topup" element={<RegisterTopup />} />
+                          <Route index element={<ActionRegisterIndex />} />
+                        </Route>
                         <Route index element={<ActionsIndex />} />
                       </Route>
                       {stakingLive && <Route path="/claim" element={<ClaimPage />} />}
